+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # v3.2.0
 - Implement PWA (Progressive Web App) offline included - see Chrome Lighthouse audits and [gatsby config for offline](./gatsby-config.js).
+- Update icon to use HTML5 shield (also in favicon) instead of Gatsby icon.
 
 # v3.1.0
 - Make the sitemap dynamic so that everytime a production build is generated in public, the xml sitemap lists all the correct pages.